qa_priority_list/requirements.txt
pyOpenSSL==17.3.0
oauth2client==4.1.2
google-api-python-client==1.7.3
google-auth==1.5.0
google-auth-httplib2==0.0.3
